Add usage examples
