+++ cli/cli.py
@@ -1,14 +1,9 @@
-import os
-import subprocess
-
 import click
-from rich.console import Console
 
-from cli.plan_executor import PlanExecutor
-from cli.constants import CHEAP_MODEL, DEFAULT_MODEL
-from cli.status_indicator import StatusIndicator
-from cli.task_runner import TaskRunner
-from cli.util import pull_branch_changes
+from cli.commands.edit import edit
+from cli.commands.plan import plan
+from cli.commands.task import task
+from cli.constants import DEFAULT_MODEL
 
 
 @click.group()
@@ -22,7 +17,24 @@ from cli.util import pull_branch_changes
 @click.option('--debug', is_flag=True, default=False, help='Display debug information.')
 @click.pass_context
 def main(ctx, wait, repo, spinner, quiet, model, branch, sync, debug):
-    """PR Pilot CLI - https://docs.pr-pilot.ai"""
+    """PR Pilot CLI - https://docs.pr-pilot.ai
+
+    Delegate routine work to AI with confidence and predictability.
+
+    Examples:
+
+    \b
+    - 📸 Create a Bootstrap5 component based on a screenshot:
+      pilot task -o component.html --code --snap "Write a Bootstrap5 component that looks like this."
+
+    \b
+    - 🛠️ Refactor and clean up code:
+      pilot edit main.js "Break up large functions, organize the file and add comments."
+
+    \b
+    - 🔄 Interact across services and tools:
+      pilot task "Find all open Linear and Github issues labeled as 'bug' and send them to the #bugs Slack channel."
+    """
     ctx.ensure_object(dict)
     ctx.obj['wait'] = wait
     ctx.obj['repo'] = repo
@@ -32,88 +44,6 @@ def main(ctx, wait, repo, spinner, quiet, model, branch, sync, debug):
     ctx.obj['branch'] = branch
     ctx.obj['sync'] = sync
     ctx.obj['debug'] = debug
-
-
-@click.command()
-@click.option('--snap', is_flag=True, help='Select a portion of your screen to add as an image to the task.')
-@click.option('--cheap', is_flag=True, default=False, help=f'Use the cheapest GPT model ({CHEAP_MODEL})')
-@click.option('--code', is_flag=True, default=False, help='Optimize prompt and settings for generating code')
-@click.option('--file', '-f', type=click.Path(exists=True), help='Generate prompt from a template file.')
-@click.option('--direct', is_flag=True, default=False,
-              help='Do not feed the rendered template as a prompt into PR Pilot, but render it directly as output.')
-@click.option('--output', '-o', type=click.Path(exists=False), help='Output file for the result.')
-@click.argument('prompt', nargs=-1)
-@click.pass_context
-def task(ctx, snap, cheap, code, file, direct, output, prompt):
-    """Create a new task for PR Pilot"""
-    console = Console()
-    show_spinner = ctx.obj['spinner'] and not ctx.obj['quiet']
-    status_indicator = StatusIndicator(spinner=show_spinner, messages=not ctx.obj['quiet'], console=console)
-
-    try:
-        if ctx.obj['sync'] and not ctx.obj['branch']:
-            # Get current branch from git
-            ctx.obj['branch'] = os.popen('git rev-parse --abbrev-ref HEAD').read().strip()
-
-        runner = TaskRunner(status_indicator)
-        runner.run_task(ctx.obj['wait'], ctx.obj['repo'], snap, None, ctx.obj['quiet'], cheap, code, file, direct, output, ctx.obj['model'], ctx.obj['debug'], prompt, branch=ctx.obj['branch'])
-        if ctx.obj['sync']:
-            pull_branch_changes(status_indicator, console, ctx.obj['branch'], ctx.obj['debug'])
-
-    except Exception as e:
-        status_indicator.fail()
-        console.print(f"[bold red]An error occurred:[/bold red] {type(e)} {str(e)}")
-    finally:
-        status_indicator.stop()
-
-
-@click.command()
-@click.argument('file_path', type=click.Path(exists=True))
-@click.argument('prompt')
-@click.pass_context
-def edit(ctx, file_path, prompt):
-    """Let PR Pilot edit a file for you."""
-    console = Console()
-    show_spinner = ctx.obj['spinner'] and not ctx.obj['quiet']
-    status_indicator = StatusIndicator(spinner=show_spinner, messages=not ctx.obj['quiet'], console=console)
-
-    try:
-        if ctx.obj['sync'] and not ctx.obj['branch']:
-            # Get current branch from git
-            ctx.obj['branch'] = os.popen('git rev-parse --abbrev-ref HEAD').read().strip()
-
-        runner = TaskRunner(status_indicator)
-        runner.run_task(ctx.obj['wait'], ctx.obj['repo'], None, file_path, ctx.obj['quiet'], False, False, None, False, None, ctx.obj['model'], ctx.obj['debug'], prompt, branch=ctx.obj['branch'])
-        if ctx.obj['sync']:
-            pull_branch_changes(status_indicator, console, ctx.obj['branch'], ctx.obj['debug'])
-
-    except Exception as e:
-        status_indicator.fail()
-        console.print(f"[bold red]An error occurred:[/bold red] {type(e)} {str(e)}")
-    finally:
-        status_indicator.stop()
-
-
-@click.command()
-@click.argument('file_path', type=click.Path(exists=True))
-@click.pass_context
-def plan(ctx, file_path):
-    """Let PR Pilot execute a plan for you."""
-    console = Console()
-    show_spinner = ctx.obj['spinner'] and not ctx.obj['quiet']
-    status_indicator = StatusIndicator(spinner=show_spinner, messages=not ctx.obj['quiet'], console=console)
-
-    try:
-        runner = PlanExecutor(file_path, status_indicator)
-        runner.run(ctx.obj['wait'], ctx.obj['repo'], ctx.obj['quiet'], ctx.obj['model'], ctx.obj['debug'])
-        if ctx.obj['sync']:
-            pull_branch_changes(status_indicator, console, ctx.obj['branch'], ctx.obj['debug'])
-
-    except Exception as e:
-        status_indicator.fail()
-        console.print(f"[bold red]An error occurred:[/bold red] {type(e)} {str(e)}")
-    finally:
-        status_indicator.stop()
 
 
 main.add_command(task)
